+++ src/main/java/fr/buildtheearth/skywindow/brigadier/context/SuggestionContext.java
@@ -1,0 +1,18 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
+package fr.buildtheearth.skywindow.brigadier.context;
+
+import fr.buildtheearth.skywindow.brigadier.tree.CommandNode;
+
+public class SuggestionContext<S> {
+    public final CommandContextBuilder<S> context;
+    public final CommandNode<S> parent;
+    public final int startPos;
+
+    public SuggestionContext(CommandContextBuilder<S> context, CommandNode<S> parent, int startPos) {
+        this.context = context;
+        this.parent = parent;
+        this.startPos = startPos;
+    }
+}